+++ src/functionalTest/kotlin/com/cmgapps/license/LicensePluginJavaShould.kt
@@ -55,12 +55,12 @@ class LicensePluginJavaShould {
 
         gradleRunner = GradleRunner.create()
             .withProjectDir(testProjectDir.toFile())
-            .withArguments(":licenseReport")
+            .withArguments(":licenseReport", "--info", "--stacktrace")
             .withPluginClasspath()
     }
 
     @ParameterizedTest(name = "${ParameterizedTest.DISPLAY_NAME_PLACEHOLDER} - Gradle Version = {0}")
-    @ValueSource(strings = ["6.8", "6.9", "7.0", "7.1", "7.2", "7.3", "7.4", "7.5", "7.6"])
+    @ValueSource(strings = ["7.2", "7.3", "7.4", "7.5", "7.6"])
     fun `apply Licenses plugin to various Gradle versions`(version: String) {
         val result = gradleRunner
             .withGradleVersion(version)